+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - refactoring the whole db layer
 - add h2 in memory db for tests
 - unit and integration tests
+- coverage diff can now go back arbitrary days (../coverage/diff/<project-name>/days/<day-to-go-back>)
 
 ## [0.7.1] - 2016-01-23
 ### Added